get_devices_from_authfile: ensure n_devs is initialized
If there's no matching user then n_devs is never initialized but we
return 1 to indicate success. A later access to n_devs will result in an
undefined read.

Note: This is currently not an issue as the single caller correctly
initializes n_devs to 0 but it affected our fuzzing process and it
future-proofs possible future callers.