pam_sm_authenticate: use goto done in two more places
It's confusing that these are the only places where return is used.

This changes the behavior if alwaysok is true and either malloc() or
pam_get_user() fails. Previously it would reject logins, with this
change it allows them. However this shouldn't affect many setups as
alwaysok is most likely not used.