util: filter authenticators based on selected options
If the user specifies PIN or user verification, only proceed with CTAP2
authenticators that support these modes. Moreover, if user verification
is disabled and unsupported by the authenticator, omit the option
instead of explicitly specifying it (as per spec).

This can be simplified when we can enforce libfido2 1.7.0 or greater
using the fido_dev_{supports,has}_{pin,uv} family of functions.