POC: Allow SIGINT while waiting for fido
This is a bad idea. This *might* work for sudo, since the sudo process
blocks SIGINT and this change overrides that temporarily. However:

a) There is no guarantee the pam-u2f plugin is being invoked by sudo
b) Even when we are running in sudo, I'm guessing sudo is not expecting
   a plugin to change the signal handlers, which invites problems

Even if we wanted to do something like this, the change is incomplete
since no cleanup of the pam-u2f resources is done when a SIGINT happens.