tests: adapt to debug_file= removal
cfg_t::debug_file is no longer assigned in tests.  Nothing else is referencing
the field, so now we can remove it too.

This commit introduces a mock syslog implementation that hijacks the
logging on stderr.  The intention is to avoid spamming the system
log when the unit tests are compiled with Automake.

The approach seems to be effective on all tested systems, but it is not clear
if relies on a well defined behaviour, that will work on *all* systems.  We
accept the risk for now, since this is only affecting the tests, only for
Automake, and since the most likely adverse outcome is syslog spamming.

The CMake build, which will eventually replace Automake, is not affected:
in such context the unit tests are linked against a dedicated build of the
tested code, compiled with the PAM_U2F_TESTING macro defined.