Adapt to new logging.h interface
Replace all debug_dbg() calls with LOG(), using an appropriate log level.

Erase the debug_dbg() as nothing is relying on it anymore.

The new LOG() macro does not depend on the configuration, so the configuration
object (of type cfg_t) does not need to be passed around as parameter.