contrib: Add git hooks
contrib/hooks now contains a collection of git hooks that the developer can use
to get faster feedback.

contrib/hooks-setup is a script that allows the developer to do a quick set up
of the provided hook. It is intended to be executed right after cloning the repo.

The contrib/hooks/commit-msg hook checks the commit message to ensure
that the git trailers in use are allowed, that is they are listed in
contrib/allowed-trailers.txt.